% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DGCclassifier.R
\name{DGCClassifier}
\alias{DGCClassifier}
\title{Fit the Gaussian copula classification model}
\usage{
DGCClassifier(x, cl, maps = TRUE, r, cdfs = "empirical", ...)
}
\arguments{
\item{x}{A discrete dataset. Assumed to be all numerical-discrete-valued if \code{maps = FALSE}.}

\item{cl}{The class labels of the dataset}

\item{maps}{Logical or a list of vectors of length \code{ncol(x)}. If FALSE, \code{x} must be a mapped dataset (i.e. all discrete values are numerical). If TRUE, all values will be mapped to discrete values in [0,1] using \code{sort(unique(x[,j]))}. If a list of vectors of length \code{ncol(x)}, values will be mapped to discrete values in [0,1] accoording to the provided list. See \code{\link[DGCPCA]{discrete_mapping}}.}

\item{r}{The number of factors in the correlation matrix approximation.}

\item{cdfs}{The marginal cdf estimations used. Possible values are "empirical" (use empirical cdfs for marginal cdfs), "multinomial-dirichlet" (estimated marginal cdfs in each class shrinks to the overall marginal cdfs using multinomial-dirichlet model), or a list of estimated cdfs of class \code{stepfun}. If a list, the data are assumed to be mapped to all numerical valued data already and the support in each margin should match the support (knots) of the cdfs provided. Each item in the list is a list of marginal cdfs for class k, and the item name should match the class name in \code{cl}.}

\item{...}{Other parameters in \code{\link[DGCPCA]{DGCFit}}.}
}
\description{
Fit the Gaussian copula classification model
}
