% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DGCclassifier.R
\name{get_bounds}
\alias{get_bounds}
\title{Get the lower and upper bound for a discrete observation}
\usage{
get_bounds(x, pmfs, cdfs)
}
\description{
Get the lower and upper bound for a discrete observation
}
