% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DGCclassifier.R
\name{gibbs_yz}
\alias{gibbs_yz}
\title{Sample y, z pairs, y is multivariate truncated normal.}
\usage{
gibbs_yz(init_y, lbounds, ubounds, rotmat, lambdas, sig, burnin, num_sample)
}
\description{
Sample y, z pairs, y is multivariate truncated normal.
}
