% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DGCclassifier.R
\name{predict.DGCClassifier}
\alias{predict.DGCClassifier}
\title{S3 prediction method for DGCClassifier}
\usage{
\method{predict}{DGCClassifier}(
  x,
  newdata,
  method = "ratio",
  base.class = "average",
  knn = 10,
  gr = T,
  hess = "explicit",
  burnin = 300,
  num_sample = 3000,
  ...
)
}
\arguments{
\item{x}{A DGCClassifier object}

\item{newdata}{The newdata to be classified. If missing, the training data will be predicted.}

\item{method}{method used to estimate likelihood ratio.}

\item{base.class}{The base class used if method = "ratio". Possible values are "max_margin", "average" or a fixed class.}

\item{knn}{The number of nearest neighbors to search if method = "entropy".}

\item{gr}{Logical. If FALSE, the gradient is approximated use finite differentiation.}

\item{hess}{The method to calculate the log determinant if the hessian matrix. "explicit" uses explicit formula and "approx" uses hessian approximated by finite differentiation in BFGS algorithm.}

\item{num_sample}{The number of samples sampled in Gibbs sampling.}

\item{burin}{The number if burn-in period in Gibbs sampling.}
}
\value{
A list with a vector of predicted class and a matrix of predicted posterior probability.
}
\description{
S3 prediction method for DGCClassifier
}
